+++ cstar/cli/cli.py
@@ -1,9 +1,9 @@
 import typer
 
 from cstar.cli.blueprint import app as app_blueprint
+from cstar.cli.common import common_callback
 from cstar.cli.environment import app as app_env
 from cstar.cli.template import app as app_template
-from cstar.cli.version import common_callback
 from cstar.cli.workplan import app as app_workplan
